refactor(android): label font
